+++ chunky/src/java/se/llbit/chunky/renderer/scene/renderbuffer/RenderTile.java
@@ -10,7 +10,7 @@ public interface RenderTile {
      * @param color Color is returned in this Vector
      * @return Number of samples
      */
-    int getColor(long x, long y, Vector3 color);
+    int getColor(int x, int y, Vector3 color);
 
     /**
      * Merge samples into a pixel.
@@ -21,7 +21,7 @@ public interface RenderTile {
      * @param b Blue component
      * @param s Number of samples
      */
-    void mergeColor(long x, long y, double r, double g, double b, int s);
+    void mergeColor(int x, int y, double r, double g, double b, int s);
 
     /**
      * Merge samples into a pixel.
@@ -30,7 +30,7 @@ public interface RenderTile {
      * @param color Color
      * @param s     Number of samples
      */
-    default void mergeColor(long x, long y, Vector3 color, int s) {
+    default void mergeColor(int x, int y, Vector3 color, int s) {
         mergeColor(x, y, color.x, color.y, color.z, s);
     }
 
@@ -43,39 +43,39 @@ public interface RenderTile {
      * @param b Blue component
      * @param s Number of samples
      */
-    void setPixel(long x, long y, double r, double g, double b, int s);
+    void setPixel(int x, int y, double r, double g, double b, int s);
 
     /**
      * Get the buffer coordinates of a pixel.
      * @param x Tile x
      * @return  Buffer x
      */
-    long getBufferX(long x);
+    int getBufferX(int x);
 
     /**
      * Get the buffer coordinates of a pixel.
      * @param y Tile y
      * @return  Buffer y
      */
-    long getBufferY(long y);
+    int getBufferY(int y);
 
     /**
      * Get the width of the buffer.
      */
-    long getBufferWidth();
+    int getBufferWidth();
 
     /**
      * Get the height of the buffer.
      */
-    long getBufferHeight();
+    int getBufferHeight();
 
     /**
      * Get the width of the tile.
      */
-    long getTileWidth();
+    int getTileWidth();
 
     /**
      * Get the height of the tile.
      */
-    long getTileHeight();
+    int getTileHeight();
 }
